refactor/data_packaging: update import paths for models
